Adding in drifters depending on Gaussian
Set up so that drifters will be seeded using a 2D Gaussian distribution around where each package was found. Then, simulations will be run starting every 4 hours in the 48 hours before the day the package was found. A Gaussian distribution is used on this timing to set the number of drifters used in each simulation.